chore: remove grunt & update dependencies
