+++ .eslintrc
@@ -1,9 +1,6 @@
 {
-  "extends": "plugin:bpmn-io/es6",
-  "env": {
-    "browser": true
-  },
-  "globals": {
-    "Promise": true
-  }
+  "extends": [
+    "plugin:bpmn-io/node",
+    "plugin:bpmn-io/browser"
+  ]
 }
