+++ package.json
@@ -3,10 +3,15 @@
   "version": "0.0.1",
   "description": "integration tests for bpmn-js",
   "scripts": {
-    "test": "grunt test",
-    "dev": "grunt auto-test",
-    "all": "npm run lint && grunt",
-    "lint": "eslint ."
+    "test": "run-s test:all",
+    "test:all": "npm run mocha -- test/spec/base/*.js test/spec/modeling/*.js test/spec/miwg/*.js",
+    "test:miwg": "npm run mocha -- test/spec/miwg/*.js",
+    "dev": "npm run test:all -- --watch",
+    "all": "run-s lint clean build test",
+    "lint": "eslint .",
+    "mocha": "mocha --reporter spec -r test/expect.js",
+    "clean": "rimraf dist tmp/integration",
+    "build": "rollup -c"
   },
   "main": "lib/main.js",
   "repository": {
@@ -37,34 +42,27 @@
     "bpmn-js": "./vendor/bpmn.js"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^22.0.2",
-    "@rollup/plugin-node-resolve": "^14.1.0",
-    "async": "^2.6.1",
+    "@rollup/plugin-commonjs": "^23.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.0",
+    "async": "^3.2.4",
     "chai": "^4.1.2",
     "debug": "^4.3.4",
     "eslint": "^8.23.0",
-    "eslint-plugin-bpmn-io": "^0.14.1",
-    "glob": "^7.1.2",
-    "grunt": "^1.0.3",
-    "grunt-cli": "^1.2.0",
-    "grunt-contrib-clean": "^2.0.0",
-    "grunt-contrib-watch": "^1.0.0",
-    "grunt-mocha-test": "^0.13.3",
-    "grunt-rollup": "^12.0.0",
-    "load-grunt-tasks": "^4.0.0",
-    "mkdirp": "^0.5.1",
+    "eslint-plugin-bpmn-io": "^0.16.0",
+    "glob": "^8.0.3",
+    "mkdirp": "^1.0.4",
     "mocha": "^10.0.0",
+    "npm-run-all": "^4.1.5",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.79.1",
     "semver": "^7.3.7",
-    "xsd-schema-validator": "^0.6.0"
-  },
-  "optionalDependencies": {
-    "node-resemble": "^1.1.1"
+    "xsd-schema-validator": "^0.7.0"
   },
   "dependencies": {
     "bpmn-js": "^10.2.0",
     "bpmn-js-cli": "^2.2.1",
     "bpmn-moddle": "^8.0.0",
-    "min-dash": "^3.0.0",
-    "puppeteer": "^17.0.0"
+    "min-dash": "^4.0.0",
+    "puppeteer": "^18.2.1"
   }
 }
